requirements.txt
PyQt5
opencv-python
scenedetect
python-dotenv
openai
pandas
openpyxl